+++ .gitignore
@@ -4,7 +4,7 @@ build/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
 !**/src/test/**/build/
-application-secret.yml
+.yml
 
 ### STS ###
 .apt_generated
