+++ build.gradle
@@ -58,6 +58,8 @@ dependencies {
 	testImplementation 'org.springframework.batch:spring-batch-test'
 	// Redis
 	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+	// https://mvnrepository.com/artifact/com.googlecode.json-simple/json-simple
+	implementation group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1'
 }
 
 dependencyManagement {
