+++ src/main/java/spharos/payment/domain/enumPackage/PaymentStatus.java
@@ -1,7 +1,7 @@
 package spharos.payment.domain.enumPackage;
 
 
-import spharos.global.config.CodeValue;
+import spharos.payment.global.config.CodeValue;
 
 public enum PaymentStatus implements CodeValue {
     WAIT("0","결제대기"),
