+++ src/main/java/spharos/payment/global/common/response/ErrorResponse.java
@@ -1,5 +1,4 @@
-package spharos.global.exception;
-
+package spharos.payment.global.common.response;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -9,20 +8,19 @@ import org.springframework.http.ResponseEntity;
 @Builder
 public class ErrorResponse {
 
-    private final int status;
-    private final String error;
-    private final String code;
+    private final boolean success;
+    private final int code;
     private final String message;
 
     public static ResponseEntity<ErrorResponse> toResponseEntity(ResponseCode responseCode) {
         return ResponseEntity
                 .status(responseCode.getHttpStatus())
                 .body(ErrorResponse.builder()
-                        .status(responseCode.getHttpStatus().value())
-                        .error(responseCode.getHttpStatus().name())
-                        .code(responseCode.name())
-                        .message(responseCode.getDetail())
+                        .success(false)
+                        .code(responseCode.getCode())
+                        .message(responseCode.getMessage())
                         .build()
                 );
     }
+
 }
