+++ src/main/java/spharos/payment/global/config/redis/RedisConfig.java
@@ -1,4 +1,4 @@
-package spharos.global.config.redis;
+package spharos.payment.global.config.redis;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
