+++ src/main/java/spharos/payment/global/config/toss/TossPaymentConfig.java
@@ -1,8 +1,11 @@
-package spharos.global.config.toss;
+package spharos.payment.global.config.toss;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 
 @Configuration
 @Getter
