+++ src/main/java/spharos/payment/global/exception/CustomException.java
@@ -1,8 +1,8 @@
-package spharos.global.exception;
+package spharos.payment.global.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import spharos.global.common.response.ResponseCode;
+import spharos.payment.global.common.response.ResponseCode;
 
 @Getter
 @AllArgsConstructor
