+++ src/main/java/spharos/payment/global/exception/ErrorResponse.java
@@ -1,4 +1,5 @@
-package spharos.global.common.response;
+package spharos.payment.global.exception;
+
 
 import lombok.Builder;
 import lombok.Getter;
@@ -8,19 +9,20 @@ import org.springframework.http.ResponseEntity;
 @Builder
 public class ErrorResponse {
 
-    private final boolean success;
-    private final int code;
+    private final int status;
+    private final String error;
+    private final String code;
     private final String message;
 
     public static ResponseEntity<ErrorResponse> toResponseEntity(ResponseCode responseCode) {
         return ResponseEntity
                 .status(responseCode.getHttpStatus())
                 .body(ErrorResponse.builder()
-                        .success(false)
-                        .code(responseCode.getCode())
-                        .message(responseCode.getMessage())
+                        .status(responseCode.getHttpStatus().value())
+                        .error(responseCode.getHttpStatus().name())
+                        .code(responseCode.name())
+                        .message(responseCode.getDetail())
                         .build()
                 );
     }
-
 }
