+++ src/main/java/spharos/payment/global/exception/GlobalExceptionController.java
@@ -1,10 +1,10 @@
-package spharos.global.exception;
+package spharos.payment.global.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import spharos.global.common.response.ErrorResponse;
+import spharos.payment.global.common.response.ErrorResponse;
 
 @Slf4j
 @RestControllerAdvice
