+++ src/main/java/spharos/payment/global/exception/ResponseCode.java
@@ -1,4 +1,4 @@
-package spharos.global.exception;
+package spharos.payment.global.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
